+++ modules/module-mysql/src/replication/MySQLConnectionManagerFactory.ts
@@ -4,23 +4,28 @@ import { MySQLConnectionManager } from './MySQLConnectionManager.js';
 import { ResolvedConnectionConfig } from '../types/types.js';
 
 export class MySQLConnectionManagerFactory {
-  private readonly connectionManagers: MySQLConnectionManager[];
+  private readonly connectionManagers = new Set<MySQLConnectionManager>();
   public readonly connectionConfig: ResolvedConnectionConfig;
 
   constructor(connectionConfig: ResolvedConnectionConfig) {
     this.connectionConfig = connectionConfig;
-    this.connectionManagers = [];
   }
 
   create(poolOptions: mysql.PoolOptions) {
     const manager = new MySQLConnectionManager(this.connectionConfig, poolOptions);
-    this.connectionManagers.push(manager);
+    this.connectionManagers.add(manager);
+
+    manager.registerListener({
+      onEnded: () => {
+        this.connectionManagers.delete(manager);
+      }
+    });
     return manager;
   }
 
   async shutdown() {
     logger.info('Shutting down MySQL connection Managers...');
-    for (const manager of this.connectionManagers) {
+    for (const manager of [...this.connectionManagers]) {
       await manager.end();
     }
     logger.info('MySQL connection Managers shutdown completed.');