+++ service/CHANGELOG.md
@@ -1,5 +1,24 @@
 # @powersync/service-image
 
+## 1.16.2
+
+### Patch Changes
+
+- d889219: Fix memory leaks when retrying replication after errors.
+- 7eb7957: [Postgres] Remove usage of pg_logical_slot_peek_binary_changes due to performance issues in some cases
+- Updated dependencies [b364581]
+- Updated dependencies [d889219]
+- Updated dependencies [0ace0d3]
+- Updated dependencies [7eb7957]
+- Updated dependencies [b364581]
+  - @powersync/service-module-mongodb-storage@0.12.11
+  - @powersync/service-core@1.16.2
+  - @powersync/service-module-postgres-storage@0.10.11
+  - @powersync/service-module-postgres@0.16.11
+  - @powersync/service-module-mongodb@0.12.11
+  - @powersync/service-module-mysql@0.9.11
+  - @powersync/service-module-core@0.2.11
+
 ## 1.16.1
 
 ### Patch Changes